improve snippet
git-svn-id: http://dev.inghist.nl/svn/BiografischPortaal/bioport_repository/trunk@22420 a13ad8ac-ce12-0410-a602-c8c22c8f6e20

+++ bioport_repository/tests/common_testcase.py
@@ -60,7 +60,8 @@ class CommonTestCase(unittest.TestCase):
         
         # remove also all data from the database
         self.repo.db.metadata.drop_all()
-        shutil.rmtree(IMAGES_CACHE_LOCAL)
+        if os.path.exists(IMAGES_CACHE_LOCAL):
+	        shutil.rmtree(IMAGES_CACHE_LOCAL)
         #self.repo.db.namenindex.db.metadata.drop_all()
         return
